project finished
textured are madeand render image is in images folder.
The complete spreadsheet is the spreadsheet file named Maya Camera Project Sheet
Render image is not as good as I wanted it to be, but that image is as good of an image as my laptop was willing to give me.